rafs: introduce is_v6()
Introduce helper is_v6() and move generic code from v5.rs into
metadata/mod.rs.
